Refactor + soundness of the membership list
- configuration creep now goes to a separate struct Config.
- each node was putting itself in its membership list upon listening to
a gossip from a peer that contained this node. Having self in the
membership list (so long as the heartbeat/suspicion are properly merged)
is not too bad, but it also doesn't yield much. Instead, having the
recipient always handle the pinger as a member not to be found in the
gossip is a cleaner way to handle this, so that's what we do now.