Remove test runner
I had started out writing a small DSL for specifying test scenarios and
an interpreter that would run the scenario (e.g., spawn k nodes, wait,
kill 1, wait, then invoke script foo with the logs for analysis). This
is not needed, and the little code duplication I'd have avoided in
the tests comes at great complexity in the runner.