core: Also store exponentially weighted moving averages and stds
Instead of keeping all the values in a window of inter arrival times and
then creating a statrs::distribution::Normal upon every tick, we can try
to use exponentially weighted moving averages and variances, and then
use the closed form formula for CDF of the normal.

    mu_new = (1 - alpha) * mu_old + alpha * curval
    var_new = (1 - alpha) * var_old + alpha * (curval - mu_old) * (curval - mu_new)

Here, alpha controls the size of the window, and in general, the number
of terms contributing significantly to the computation of new values is
given by

    1 / alpha

For example, for alpha = 0.1, we have

    mu_new = 0.9 * mu_old + 0.1 * curval
    var_new = 0.9 * var_old + 0.1 * (curval - mu_old) * (curval - mu_new)
    num_terms = 1 / 0.1 = 10

So roughly a window size of 10. This seems to be working pretty well,
next step is to scrap the Normal object and use the the mean and
variance to directly compute the CDF.